use BufferSubData for uploading rf data on NVIDIA/AMD
It may seem counter-intuitive and goes against all of NVIDIA's
presentations and other documentation but this is the fastest way
to upload the data.

Unfourtanetely this doesn't work on Intel for some reason so it
gets a seperate path. The seperate path code is necessary anyways
because we want to add CUDA support for NVIDIA.