ui: don't shoehorn every listing through the same loop
Technically I could keep the modifiable elements grouped and loop
over them but I want to be able to add buttons and they will need
a hover_t but not the other parameters. This may require slightly
more code but it keeps things clear.

This also fixes the issue with scrolling on a selected field and
scrolling the dB value causing a recompute of the output.