core: delete all multi transducer handling code
This will be reintroduced later by using a seperate UBO for each
transducer. Not only will this be more general, but it also
significantly simplifies the code. The only thing necessary
besides providing an api for submitting multiple UBOs is adding a
UBO field to the beamform compute work items.

The code is now limited to handling a maximum of 256 transmit
events for a single transducer. This should be more than
suffiecent for any arrays we manufacture in the lab.