build.sh: built raylib's bundled glfw separately
raylib is not-architected to have multiple OpenGL contexts so we
need to do it behind their back. We can still use the bundled GLFW
but we have to tell raylib that it is external.

This also makes it so that touching the build.sh will cause the
built external libraries to rebuild.