intrinsics: avoid UB when using CLZ/CTZ
Despite using the compiler builtins and the underlying assembly
instruction not caring if the value is 0 it is still undefined
behaviour in C to call these functions with an argument of 0.

Add a little check to workaround this which will be optimized out
and any optimization level other that 0. See: https://godbolt.org/z/aoGx7PP5K