core: remove oversized raw_data_ssbo
I don't think it was ever being used correctly and now it is
pointless due to us calling glFinish() before trying to move onto
another frame. If at some point the data is not being loaded by
the worker thread then the fences could be reintroduced but I
don't really see the point.