core: integrate renderdoc frame capturing in debug build
There is still a caveat that renderdoc is unable to capture only a
single OpenGL context when multiple are present even when you tell
it which one to track. This will at least capture all dispatch
compute calls but they will interleaved with a bunch of garbage
from the other context that we don't really care about. Also this
sometimes misses the raw and decoded data buffers but its better
than what we had before.

closes: #23