ui: fix half baked scratch variable idea
The active input state Variable needs to be a pointer. To avoid
clobbering the active variable when it is pointing to the scratch
variable we need a second scratch variable. They can be swapped
everytime we assign the active variable. It doesn't actually
matter if the hot variable is currently pointing to one or the
other when we swap, just that we know that the scratch variable is
not the currently active variable.

Also move flags into the Variable so that we can determine if any
arbitrary variable should accept text input or not.