core/lib: properly check for timeout from wait functions
Also just use the Wait functions from Synchronization.lib on
Windows. They don't seem to have any major issues.

For now, the only supported timeout for the library push functions
on w32 for is 0 to poll. The cross process synchronization will
need a rework to be supported on w32. Its fine to just check the
return value on the functions for success.