das: make cubic equation read closer to linked reference
ours still looks different simply because we are doing 2
interpolations in parallel (one for the real data (in-phase) and
one for the complex data (quadrature))

also this is consistently faster by ~8ms per frame on my desktop.
I'm guessing the driver is better able to optimize this when it is
properly stated as a series of matrix multiplications instead of
decomposed dot products.