core/lib: coalesce buffer uploads that haven't made it to the GPU
If a caller overwrites a buffer that hasn't been uploaded to the
GPU yet we don't need to make a new work item. The library code is
already holding the lock so we know that beamformer can't upload
the buffer until we release it so overwrite the data and only push
the work item if the dirty region flag is not set.