shaders/decode: improve gpu cache behaviour
for a single work group we do not want data from more than 1
channel. Assuming a worse case sample count of 4096 samples per
transmit we can fit 16 transmits in a 128KB L1 cache (pretty
common even on low end modern cards). Therefore we want dispatches
in the transmit dimension to be 16.

In the samples dimension the dispatch count is a little less
obvious but we want it to be at least 2 on nVidia or 4 on AMD. AMD
typically reports an L0 (per work group) Cache which we can maybe
use to make a better decision there. But on my (AMD) card anything
above 4 performed essentially the same. Choosing 4 is fine for an
nVidia card too but only 32 items will actually get scheduled on
the same SM.

This results in a ~24% throughput increase on my system for
FORCES/HERCULES. The decode shader itself is now more than twice
as fast.