core: allow non power of 2 images to be created
These will result in lots of wasted compute and is not recommended
but will keep the result from a library perspective expected.
Before this the library would try to read the wrong amount of data
and lock up/timeout waiting for data that would never come.

For optimal compute usage images should be a power of 2 multiple
of DAS_LOCAL_SIZE in each dimension.