ui: solve VariableGroup member ordering mystery
this was left over from when I thought it was a good idea to try
and make UIView an implicit superset of VariableGroup ala C++
inheretince (there used to be a VariableGroup at the base of
UIView). This continued to work when I removed that because there
was an overlap in Variable between the VariableGroup.first and
UIView.child.

DO NOT TRY TO DO MAGICAL THINGS! IT IS STUPID!