ui: no longer destroy self on hot reload
while it was a cool trick to support some structs changing in
definition some GPU resources were being leaked. It is also more
useful if ui layout remained fixed when working on interaction
code or drawing code. w.r.t the static data comment we do not
currently have any and if we did it would go in the static portion
of the executable not the hot reloadable portion.