core/build: go all in on ubsan/asan
ubsan/asan are now always enabled in debug builds (they can still
be optionally enabled in release build).

For asan purposes freelist members are now poisoned when pushed
and unpoisoned when popped. Arenas are fully poisoned at
appropriate places but this has some drawbacks, particularly for
arena_stream(). It will catch anything in a frame up until a call
to arena_stream(). That part could be worked on later as
necessary.

This also fixes debug build with GCC (specifically with ld as the
linker) which was broken since 45daa35.

Finally, a bunch of memory was moved off the stack and pushed into
the arena to improve locality.