core: shave 40% off DAS runtime
for FORCES and HERCULES the speed was improved by beamforming
receive channel by receive channel after resetting the order of
the rf data to the order it was originally received in. This order
is more cache friendly. VLS/TPW most likely perform better in the
old shader because focal vectors can be better utilized from
cached texture memory but it wasn't explored in detail.

Other options making use of GPU shared memory all performed way
worse than the existing code.

New shader is currently auto selected if we can determine from the
rest of the parameters that it is supported.