shaders/das: early out when channel apodization is 0
This is most obviously beneficial for FORCES in the DASFast path
where it is clear that it skips the entire computation for that
thread. When you consider how the voxels are clustered it is
obvious that a large portion of warps can be immediately retired
when there is a receive f-number applied (think on individual
channel level where a large portion of the image is completely
black).

This also gives a performance boost to HERCULES and the RCA
methods but to a lesser extent since it only avoids rf data
sampling in those threads at the expense thread divergence.

On its own this change seems trivial but it brings to light that
the correct way to do DAS is via indirect dispatch from another
compute shader. That way we won't even need to issue warps for
threads which would early out due to apodization. As I implement
it I'm sure there will be other conditions for skipping full DAS
on a voxel which will be highly beneficial.