core/lib: rework pipeline handling
There are cases such as with DASFast where the compute thread
wants to modify the pipeline on its own. However this caused
problems with the library preempting the compute thread during
operation leading to issues such as the incorrect shader being
used.

This lead to the realization that the compute thread should have
its own stable pipeline to operate on which it can setup during a
'planning' phase. During this planning phase the compute thread
can modify the parameters as necessary to achieve optimal
performance before uploading its own copy to the GPU.

This can also aid in simplifying the library API. For example the
planning phase accounts for the time delay introduced by any
applied filters.