core: bind only filled portion of rf buffer for DAS
I'm still not sure how a channel is accessing data beyond the
region that was filled and reading NaNs but it is. This prevents
part of the image from being polluted.

This only seems to happen when we demodulate after decoding which
doesn't make a whole lot of sense to me. A similiar effect can be
created by using an unsupported decimation rate. But in that case
demod first produces a buffer that won't beamform correctly but
still contains non NaN data.

Future commit will size the buffer in plan_compute_pipeline() but
when the beamformer decodes first it still needs to be oversized
to accommodate the extra samples.