build: output matlab beamformer classes
these should prevent our most common pain point when interfacing
with matlab. when we try to modify the parameters and add or
remove a field and we use an incorrect name the only error we get
is "a field doesn't match one in the struct" which is obviously
completely useless trash that only mathworks could come up with.

This will allow me to modify these structs and get a reasonable
error message in matlab when a field is no longer valid.

These classes should be directly convertible to a struct via struct(obj).